Fixed gray shapes
Fixed the gray shapes, because of their containers they were actually causing the entire screen to stretch out of bounds and in a way. Either way I added the top gray shape, and fixed the bottom one as well. All that's left for this page is nav links and responsiveness.