web design page almost finished
This page just needs the two little gray cards at the bottom. After all of that is finished, i'll have the exact template necessarily to copy and paste over for the next 2 pages. All that will leave is the last 3 pages contact, about us, and location. As well as the rest of the hover affects and breakpoints/responsiveness.